Allow using subject from CSR instead of overwriting.
In some rare cases we want to accept the Subject contained in a
certificate signing request (CSR) instead of overwriting it. Add a
command line option to preserve the Subject into the resulting
certificate. This is for advanced usage, not day to day
operations. The default is to overwrite the subject.