Allow SA.get_credential to take any URN as argument
Generalize SA.get_credential for both slices or projects by making the
argument a URN instead of forcing it to be a Slice URN. Plumb this
through to the SA implementation and split get_credential into two
parts, one for slices and the other for projects.
get_credential(Project) is a stub that yields a not implemented
error.