Communicate max slice expiration in SA get_version
Client programs have no way to no what the max slice expiration is. Add
the max expiration to the slice authority get_version response so that
clients can use the information to determine requested slice lifetimes.

Use the key '_GENI_MAX_SLICE_EXPIRATION_DAYS' and set the value
appropriately with the unit of days from the constant used in the SA.