Swap identities with CH generated id
Have the CH generate the swap id (formerly nonce) instead of
allowing the client to pass one in. This allows the CH to control
uniqueness more easily. The generated id is returned from the
call to create_swap_id(), and is later used with swap_identities()
to swap two identities.