Update CredentialStore.py
Filter DB queries to only look for un-expired slices and projects. Grabbing everything and passing it on to the GENI Portal that uses a PHP based xmlrpc_decode function causes the routine to crash there by rendering the user's GENI portal account useless and throw CH API errors. This has reduced the number of items sent as a result for the query. 
For a GENI user that trigger this issue, before the optimization, projects = 146 and slices > 52K were returned causing xmlrpc_decode (PHP) to return NULL as a crash indicator. After the filter the number has gone down to projects=19 and slices=743 and GENI Portal now works for the user again.